feat!: additional updates to support ESM
* providing Node.js args directly in binary
* updated package.json entry fiels for types and entry point definitions
* updated Yargs configuration to the one compatible with ESM
