+++ package.json
@@ -1,9 +1,13 @@
 {
   "name": "jira-miner",
-  "version": "5.0.0-beta.1",
+  "version": "5.0.0-beta.2",
   "description": "An utility for advanced queryies on top of JIRA",
   "type": "module",
   "exports": {
+    ".": {
+      "types": "./dist/index.d.mts",
+      "import": "./dist/index.mjs"
+    },
     "./*": "./dist/*.js"
   },
   "typesVersions": {
@@ -13,10 +17,8 @@
       ]
     }
   },
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
   "bin": {
-    "jira-miner": "./dist/index.js"
+    "jira-miner": "./dist/index.mjs"
   },
   "repository": {
     "type": "git",
