+++ tsconfig.json
@@ -2,7 +2,7 @@
   "extends": "@tsconfig/node16/tsconfig.json",
   "compilerOptions": {
     "target": "ES2021",
-    "module": "es2020",
+    "module": "es2022",
     "moduleResolution": "node",
     "sourceMap": true,
     "noImplicitAny": false,
